기능 수정
- 중첩 상태 스택 감소 및 제거 기능을 턴이 오면 스킵없이 바로 사라지게 수정
- 스킬 대미지 적용 수치 적용을 RoundToInt방식에서 FloorToInt 로 변경(소수점 버림)
- 공격 범위 하이라이트 표시를 맵 타일 내에 있는 범위만 표시되게 수정(rollback)

+++ Assets/Resources/Scripts/Battle/SkillAsset.cs
@@ -79,7 +79,7 @@ public abstract class SkillAsset : ScriptableObject
         var list = candidates.Where(u => u != null && u.team == Team.Player && !u.IsDead).ToList();
         if (list.Count == 0) return null;
 
-        int max = list.Max(u => Mathf.Max(0, u.Hostility));
+        float max = list.Max(u => Mathf.Max(0, u.Hostility));
         var top = list.Where(u => Mathf.Max(0, u.Hostility) == max).ToList();
 
         return top[Random.Range(0, top.Count)];
